Implemented conversion from bf16 to f8e4m3

+++ src/plugins/intel_cpu/tests/functional/shared_tests_instances/single_layer_tests/conversion.cpp
@@ -19,7 +19,8 @@ const std::vector<std::vector<ov::Shape>> shapes = {{{1, 2, 3, 4}}};
 
 const std::vector<ov::element::Type> inTypes = {
     // ov::element::f32,
-    ov::element::f16,
+    // ov::element::f16,
+    ov::element::bf16,
 };
 
 const std::vector<ov::element::Type> outTypes = {
